Move ClearRemindersBacklogTaskCog class to remind_me.py cog
